+++ tests/common/defines.hpp
@@ -38,3 +38,11 @@
             throw std::runtime_error(errmsg.str());                                                \
         }                                                                                          \
     }
+
+#if HIP_VERSION >= 60300000
+#    define HIP_HOST_ALLOC_FUNC hipExtHostAlloc
+#    define HIP_HOST_FREE_FUNC  hipFreeHost
+#else
+#    define HIP_HOST_ALLOC_FUNC hipHostMalloc
+#    define HIP_HOST_FREE_FUNC  hipHostFree
+#endif