+++ CHANGELOG.md
@@ -101,6 +101,7 @@ Full documentation for ROCprofiler-SDK is available at [Click Here](source/docs/
 ### Changes
 
 - Support `--marker-trace` on application linked against old (roctracer) ROCTx (i.e. `libroctx64.so`)
+- Replaced deprecated hipHostMalloc and hipHostFree functions with hipExtHostAlloc and hipFreeHost in when ROCm version is greater than or equal to 6.3
 
 ### Fixes